+++ test/IntegrationTest.php
@@ -49,13 +49,7 @@ class IntegrationTest extends TestCase
         $request  = new ServerRequest([], [], 'https://example.com/foo', 'GET');
         $response = new Response();
 
-        set_error_handler(function ($errno, $errstr) {
-            return false !== strstr($errstr, 'OriginalMessages');
-        }, E_USER_DEPRECATED);
-
         $app->run($request, $response);
-
-        restore_error_handler();
 
         $this->assertInstanceOf(ResponseInterface::class, $this->response);
         $this->assertEquals(404, $this->response->getStatusCode());
@@ -68,13 +62,7 @@ class IntegrationTest extends TestCase
         $request      = new ServerRequest([], [], 'https://example.com/foo', 'GET');
         $response     = new Response();
 
-        set_error_handler(function ($errno, $errstr) {
-            return false !== strstr($errstr, 'OriginalMessages');
-        }, E_USER_DEPRECATED);
-
         $app->run($request, $response);
-
-        restore_error_handler();
 
         $this->assertInstanceOf(ResponseInterface::class, $this->response);
         $this->assertEquals(404, $this->response->getStatusCode());